+++ back-end/macademia/Department.java
@@ -1,4 +1,4 @@
-package macademia.backend;
+package macademia;
 
 import java.util.Hashtable;
 import java.util.Map;
