--- back-end/macademia/backend/.gitignore
@@ -1,3 +1,0 @@
-/User.class
-/Student.class
-/Department.class
